+++ db/migrate/20140915202917_add_auto_subscribe_to_users.rb
@@ -1,0 +1,5 @@
+class AddAutoSubscribeToUsers < ActiveRecord::Migration
+  def change
+    add_column :users, :auto_subscribe, :boolean, null: false, default: true
+  end
+end